Refactor media upload to use byte array instead of stream
This commit changes the method for uploading media in the MastodonClient, MediaDefinition, IMastodonClient, and BaseHttpClient classes from using a Stream object to a byte array. This makes the handling of the media data more consistent and versatile. The update also adds some optional parameters to the OAuthUrl method in IAuthenticationClient and AuthenticationClient for better functionality.